Change folder name reducers to features
