+++ App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Provider } from 'react-redux';
-import store from './reducers/store';
+import store from './features/store';
 import AppNavigator from './navigations/AppNavigator';
 
 export default function App() {
