+++ components/shared/Toggle.js
@@ -1,6 +1,6 @@
 import React, { useRef } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { toggleFilter } from '../../reducers/postSlice';
+import { toggleFilter } from '../../features/postSlice';
 import { Animated, View, Text, Pressable, StyleSheet, Easing } from 'react-native';
 import colors from '../../theme/color';
