--- reducers/chatSlice.js
@@ -1,32 +1,0 @@
-import { createSlice } from '@reduxjs/toolkit';
-
-const initialState = {
-  isEntered: false,
-};
-
-export const chatSlice = createSlice({
-  name: 'chat',
-  initialState,
-  reducers: {
-    enterChat(state) {
-      state.isEntered = true;
-    },
-    leaveChat(state) {
-      state.isEntered = false;
-    }
-  },
-  extraReducers: {
-  }
-});
-
-const {
-  actions,
-  reducer
-} = chatSlice;
-
-export const {
-  enterChat,
-  leaveChat
-} = actions;
-
-export default reducer;
