+++ screens/ChatRoom.js
@@ -1,6 +1,6 @@
 import React, { useEffect, useState, useRef } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { enterChat, leaveChat } from '../reducers/chatSlice';
+import { enterChat, leaveChat } from '../features/chatSlice';
 import { useRoute } from '@react-navigation/native';
 
 import {
